Location and City name
