+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:usesCleartextTraffic="true">
-
+        <activity android:name=".Graph"></activity>
         <activity android:name=".Dashboard" />
         <activity
             android:name=".Initiate"
