+++ app/src/main/res/layout/activity_dashboard.xml
@@ -50,12 +50,14 @@
                     android:rowCount="2">
 
                     <LinearLayout
+                        android:id="@+id/start_temp_graph"
                         android:layout_width="0dp"
                         android:layout_height="0dp"
                         android:layout_rowWeight="1"
                         android:layout_columnWeight="1"
                         android:background="@android:color/white"
                         android:elevation="1.6dp"
+                        android:clickable="true"
                         android:orientation="horizontal">
 
                         <ImageView
@@ -90,12 +92,14 @@
                     </LinearLayout>
 
                     <LinearLayout
+                        android:id="@+id/start_humi_graph"
                         android:layout_width="0dp"
                         android:layout_height="0dp"
                         android:layout_rowWeight="1"
                         android:layout_columnWeight="1"
                         android:background="@android:color/white"
                         android:elevation="1.65dp"
+                        android:clickable="true"
                         android:orientation="horizontal">
 
                         <ImageView
@@ -130,12 +134,14 @@
                     </LinearLayout>
 
                     <LinearLayout
+                        android:id="@+id/start_wind_graph"
                         android:layout_width="0dp"
                         android:layout_height="0dp"
                         android:layout_rowWeight="1"
                         android:layout_columnWeight="1"
                         android:background="@android:color/white"
                         android:elevation="1.5dp"
+                        android:clickable="true"
                         android:orientation="horizontal">
 
                         <ImageView
@@ -171,12 +177,14 @@
                     </LinearLayout>
 
                     <LinearLayout
+                        android:id="@+id/start_rain_graph"
                         android:layout_width="0dp"
                         android:layout_height="0dp"
                         android:layout_rowWeight="1"
                         android:layout_columnWeight="1"
                         android:background="@android:color/white"
                         android:elevation="1.6dp"
+                        android:clickable="true"
                         android:orientation="horizontal">
 
                         <ImageView
@@ -241,48 +249,7 @@
             android:textSize="11.4dp"
             app:layout_constraintStart_toStartOf="@+id/current_temp"
             app:layout_constraintTop_toBottomOf="@+id/current_temp" />
-        <ScrollView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layout_constraintTop_toBottomOf="@id/cardView">
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="4500dp"
-                android:orientation="vertical"
-                android:paddingBottom="20dp">
-                <androidx.cardview.widget.CardView
-                    android:id="@+id/chartCard"
-                    android:layout_width="match_parent"
-                    android:layout_height="250dp"
-                    android:layout_marginTop="5.7dp"
-                    android:layout_marginBottom="5.7dp"
-                    android:elevation="5dp"
-                    android:layout_marginStart="11.4dp"
-                    android:layout_marginEnd="11.4dp"
-                    app:cardCornerRadius="5dp"
-                    tools:layout_editor_absoluteX="11dp">
 
-                    <LinearLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="match_parent"
-                        android:visibility="visible"
-                        android:gravity="center">
-                        <com.wang.avi.AVLoadingIndicatorView
-                            android:id="@+id/chartLoader"
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:visibility="visible"
-                            app:indicatorColor="@color/chart_grad_top"
-                            app:indicatorName="BallClipRotateMultipleIndicator" />
-                    </LinearLayout>
-                    <com.github.mikephil.charting.charts.LineChart
-                        android:id="@+id/chart"
-                        android:layout_width="match_parent"
-                        android:layout_height="match_parent" />
-                </androidx.cardview.widget.CardView>
-            </LinearLayout>
-        </ScrollView>
     </androidx.constraintlayout.widget.ConstraintLayout>
 
 </LinearLayout>
