+++ app/src/main/res/layout/activity_graph.xml
@@ -1,0 +1,65 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".Graph"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/chart_name"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textAlignment="center"
+        android:layout_marginTop="20dp"
+        android:textSize="40dp"
+        android:text="Temperature"/>
+
+    <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:layout_marginTop="50dp"
+        android:layout_marginHorizontal="7dp"
+        app:layout_constraintTop_toBottomOf="@id/cardView">
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="4500dp"
+            android:orientation="vertical"
+            android:paddingBottom="20dp">
+            <androidx.cardview.widget.CardView
+                android:id="@+id/chartCard"
+                android:layout_width="match_parent"
+                android:layout_height="250dp"
+                android:layout_marginTop="5.7dp"
+                android:layout_marginBottom="5.7dp"
+                android:elevation="5dp"
+                android:layout_marginStart="11.4dp"
+                android:layout_marginEnd="11.4dp"
+                app:cardCornerRadius="5dp"
+                tools:layout_editor_absoluteX="11dp">
+
+                <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:visibility="visible"
+                    android:gravity="center">
+                    <com.wang.avi.AVLoadingIndicatorView
+                        android:id="@+id/chartLoader"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:visibility="visible"
+                        app:indicatorColor="@color/chart_grad_top"
+                        app:indicatorName="BallClipRotateMultipleIndicator" />
+                </LinearLayout>
+                <com.github.mikephil.charting.charts.LineChart
+                    android:id="@+id/chart"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent" />
+            </androidx.cardview.widget.CardView>
+        </LinearLayout>
+    </ScrollView>
+
+</LinearLayout>
